+++ Coursework/views/pages/movieshowinfo.ejs
@@ -7,6 +7,7 @@
         <link rel="stylesheet" type="text/css" href="/css/movieshowinfo.css">
         <script src='/scripts/require.js'></script>
         <script src="/scripts/movieinfo.js"></script>
+        <script src="/scripts/mongo.js"></script>
         <!-- Linking the stylesheet for the index page -->
 </head>
